refactor(web): migrate googledrive/dropbox/playlists APIs to ReScript (Phase 3 cont.)
Three more live modules on the Axios.res params foundation. No consumer
imported a type from these (functions only), so returns are typed concretely
where they were untyped `any` before.

- Googledrive.res: getFiles -> filesListing (parentDirectory/directory/
  directories/files), getFile -> fileMeta. Nested dir/file records.
- Dropbox.res: same filesListing shape; getFile -> fileMeta whose Dropbox
  `.tag` key is bound via @as(".tag") (verified the gen renders `".tag"`);
  getTemporaryLink -> { link }. getFile's odd same-endpoint-as-getFiles call
  with a `path` param is preserved verbatim.
- Playlists.res: getPlaylists -> playlist[] (reads `.data.playlists`).

Dropped playlists.ts's `getPlaylist`: it was dead — usePlaylistQuery actually
calls getPlaylists, and the live single-playlist fetch is the inline
usePlaylists hook hitting a different endpoint.

`at`/`path`/`id`/`did` all go through axios `params` (undefined `at` dropped
by axios, matching the old `{ params: { at } }`).

Verified: rescript build, tsc -b, full vite build.
